refactor code and separate each service in DI class
